Templatify cards

+++ routes/main.go
@@ -90,7 +90,7 @@ func SpecificRoomsHandler(ctx *macaron.Context) {
 		ctx.Data["Devices"] = models.GetDevices()
 		for _, r := range simulation.Env.Rooms {
 			if r.DBRoomID == room.RoomID {
-				ctx.Data["Temp"] = r.ActualRoomTemp
+				ctx.Data["Temp"] = int64(r.ActualRoomTemp)
 				break
 			}
 		}
